Trait sui_aws_orchestrator::protocol::ProtocolMetrics

pub trait ProtocolMetrics {
    const BENCHMARK_DURATION: &'static str;
    const TOTAL_TRANSACTIONS: &'static str;
    const LATENCY_BUCKETS: &'static str;
    const LATENCY_SUM: &'static str;
    const LATENCY_SQUARED_SUM: &'static str;

    // Required methods
    fn nodes_metrics_path<I>(&self, instances: I) -> Vec<(Instance, String)>
       where I: IntoIterator<Item = Instance>;
    fn clients_metrics_path<I>(&self, instances: I) -> Vec<(Instance, String)>
       where I: IntoIterator<Item = Instance>;

    // Provided methods
    fn nodes_metrics_command<I>(&self, instances: I) -> Vec<(Instance, String)>
       where I: IntoIterator<Item = Instance> { ... }
    fn clients_metrics_command<I>(
        &self,
        instances: I,
    ) -> Vec<(Instance, String)>
       where I: IntoIterator<Item = Instance> { ... }
}

The names of the minimum metrics exposed by the load generators that are required to compute performance.

Required Associated Constants

const BENCHMARK_DURATION: &'static str

The name of the metric reporting the total duration of the benchmark (in seconds).

const TOTAL_TRANSACTIONS: &'static str

The name of the metric reporting the total number of finalized transactions

const LATENCY_BUCKETS: &'static str

The name of the metric reporting the latency buckets.

const LATENCY_SUM: &'static str

The name of the metric reporting the sum of the end-to-end latency of all finalized transactions.

const LATENCY_SQUARED_SUM: &'static str

The name of the metric reporting the square of the sum of the end-to-end latency of all finalized transactions.

Required Methods

fn nodes_metrics_path<I>(&self, instances: I) -> Vec<(Instance, String)>
where I: IntoIterator<Item = Instance>,

The network path where the nodes expose prometheus metrics.

fn clients_metrics_path<I>(&self, instances: I) -> Vec<(Instance, String)>
where I: IntoIterator<Item = Instance>,

The network path where the clients expose prometheus metrics.

Provided Methods

fn nodes_metrics_command<I>(&self, instances: I) -> Vec<(Instance, String)>
where I: IntoIterator<Item = Instance>,

The command to retrieve the metrics from the nodes.

fn clients_metrics_command<I>(&self, instances: I) -> Vec<(Instance, String)>
where I: IntoIterator<Item = Instance>,

The command to retrieve the metrics from the clients.

Object Safety

This trait is not object safe.